added Activity logic, sorting and map
